+++ src/command_modules/azure-cli-sql/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+Unreleased
++++++++++++++++++++
+* Adding support for SQL Transparent Data Encryption (TDE) and TDE with Bring Your Own Key
+
 2.0.12 (2017-09-22)
 +++++++++++++++++++
 * az sql server list --resource-group argument is now optional. If not specified, all sql servers in the entire subscription will be returned.